NS-72/endpoint-managers: Rename bandwidth_per endpoint -> network_bandwidth_per_endpoint
