+++ net-gateway/src/endpoints/charts/mod.rs
@@ -1,3 +1,3 @@
-pub mod bandwidth_per_endpoint;
+pub mod network_bandwidth_per_endpoint;
 pub mod network_bandwidth;
 pub mod network_graph;
